requirements.txt
fastapi==0.111.0
pydantic==2.7.1
pydantic_core==2.18.2
uvicorn==0.29.0
environ-config==21.2.0
requests==2.31.0
grpcio==1.64.1
grpcio-tools==1.64.1
googleapis-common-protos==1.63.2
grpcio_status==1.64.1
protobuf==5.27.2
elastic-apm==6.22.3
tabulate
ipython
loguru==0.7.2
pycountry==24.6.1


#Opentelemetry
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-fastapi


prometheus-client==0.21.1


factory_boy
faker
locust

#dev
mypy==1.10.1
pytest==8.2.2
mypy-protobuf==3.6.0
safety==3.2.4
bandit==1.7.9
ruff==0.5.1
pytest-cov==4.1.0
pytest-sugar==0.9.7
pytest-xdist==3.3.1
pytest-asyncio==0.23.7
httpx==0.27.0